Dispatcher v.0.4
In this version the FCFS and the Non Preemptive algorithms are already
covered. There is lack of documentation since I am using a Rapid
Application Model to code. Also, I am not focusing in reusing methods
nor components. The most important update in this version is the main
structure and the way that we are going to work.