DispatcherSimulator v.0.8
Mostly all algorithms are done, adding the Preemptive ones and Round
Robin. It has been tested only with the "process4.txt" file. This does
not represent the final version nor a stable one. It lacks of
documentation, organisation, and factorisation. It's mainly purpose is
to be used as a testing version for the algorithms implementations.