DispatcherSimulator v.0.93
Fixed a lot of bugs for the Preemptive algorithms:
-- Interruptions of Interruptions were not stored in the second process
list.
-- If a process which arrival time were less than a process whom have
interrupted another process before, interruptions were not made unless
the process being executed runs for a long or same time as the arrival
time of the process that must interrupt its execution.
-- Interruptions were made after a period of time, which modifies the
real average waiting time.
-- Overflows in some special cases.

All the code is mostly commented and explained, it's just leftover the
README file.